공지사항 삭제 취소시 삭제되는거 유효성체크 기능 추가

+++ ProjectPage/src/main/webapp/WEB-INF/views/admin/adminBoardList.jsp
@@ -95,31 +95,35 @@
    });
    
    /* 체크박스 값 배열로 저장 */
-	/*
-	var total_cnt : 배열의 index 
-	var checkArray : 다중체크박스 배열
-	체크박스를 여러개 선택하면 배열로 저장한 뒤에  data-num 속성에 배열을 넣어주고 
-	controller에서 반복문을 이용해서 num 값을 하나씩 빼서 삭제함. 
-	*/
-	$('.delete-btn').click(function() {
-		var total_cnt = 0;
-		var checkArray = new Array();
-		
-		$('.checkbox').each(function() {
-			if (this.checked) {
-				checkArray[total_cnt] = this.value;	
-				total_cnt++;
-			}
-		});
-		
-		document.getElementById('delete-btn').setAttribute('data-num', checkArray);
-		
-		if (checkArray == ''  && checkArray == 0) {
-			alert('최소한 1개는 선택해야합니다.');
-		} 
-		if(checkArray != ''  && checkArray != 0){
-			var delete1 = confirm("삭제 하시겠습니까?");
-			location.href='${pageContext.request.contextPath}/admin/adminBoardDelete.do?num='+$(this).attr('data-num');
-		}
-		});
+   /*
+   var total_cnt : 배열의 index 
+   var checkArray : 다중체크박스 배열
+   체크박스를 여러개 선택하면 배열로 저장한 뒤에  data-num 속성에 배열을 넣어주고 
+   controller에서 반복문을 이용해서 num 값을 하나씩 빼서 삭제함. 
+   */
+   $('.delete-btn').click(function() {
+      var total_cnt = 0;
+      var checkArray = new Array();
+      
+      $('.checkbox').each(function() {
+         if (this.checked) {
+            checkArray[total_cnt] = this.value;   
+            total_cnt++;
+         }
+      });
+      
+      document.getElementById('delete-btn').setAttribute('data-num', checkArray);
+      
+      if (checkArray == ''  && checkArray == 0) {
+         alert('최소한 1개는 선택해야합니다.');
+      } 
+      if(checkArray != ''  && checkArray != 0){
+         var delete1 = confirm("삭제 하시겠습니까?");
+         if(delete1 == false){
+            return false;
+         }else{
+            location.href='${pageContext.request.contextPath}/admin/adminBoardDelete.do?num='+$(this).attr('data-num');
+         }
+      }
+      });
 </script>
